refactor(studio): strip navbar and hoist training unload guard
